🛠️ refactor: move voice chunking logic to a dedicated module
